MTGFLOW/requirements.txt
 pytorch==1.7.1
 numpy==1.19.2
 torchvision==1.5
 scipy==1.6.1
 scikit-learn==0.24.1
 scikit-image==0.18.1
 matplotlib== 3.3.4
 pillow == 7.2.0
